refactoring, moving things to the PPO agent
