+++ deep_quoridor/src/agents/sb3_ppo.py
@@ -1,5 +1,8 @@
+import torch
+from gymnasium import spaces
 from pettingzoo.utils import BaseWrapper
 from sb3_contrib import MaskablePPO
+from stable_baselines3.common.torch_layers import FlattenExtractor
 
 from agents.core.agent import ActionLog, Agent, AgentRegistry
 from agents.core.trainable_agent import AbstractTrainableAgent, TrainableAgentParams
@@ -136,7 +139,7 @@ class SB3PPOAgent(AbstractTrainableAgent):
             self.fetch_model_from_wand_and_update_params()
 
             # Wrap the game to get access to action_mask method
-            self.wrapper = SB3ActionMaskWrapper(game)
+            self.wrapper = wrap_env(game)
 
             try:
                 # Find the most recent model file
@@ -172,6 +175,37 @@ class SB3PPOAgent(AbstractTrainableAgent):
         action = int(self.model.predict(observation, action_masks=action_mask, deterministic=self.deterministic)[0])
 
         return action
+
+
+def wrap_env(env):
+    env = SB3ActionMaskWrapper(env)
+    return env
+
+
+def make_env_fn(env_constructor):
+    def env_fn(**kwargs):
+        env = env_constructor(**kwargs)
+        env = wrap_env(env)
+        return env
+
+    return env_fn
+
+
+class DictFlattenExtractor(FlattenExtractor):
+    """
+    This class is necessary because the default FlattenExtractor does not work with dict spaces.
+    It just tries to call "flatten" directly which of course doesn't exist on a dict.
+    NOTE: It is important to be careful to not flatten the batch dimension (first one).
+    """
+
+    def __init__(self, observation_space: spaces.Box):
+        super().__init__(observation_space)
+
+    def forward(self, obs: dict) -> torch.Tensor:
+        thobs = torch.tensor([], device=list(obs.values())[0].device)
+        for v in obs.values():
+            thobs = torch.cat((thobs, torch.tensor(v).flatten(start_dim=1)), dim=1)
+        return thobs
 
 
 # Register the agent with the registry
